Converted ADC output back to int type, with bit mask to prevent overflow. Added switch case to convert bit values to ADC channels in ADC return function. Made ADC global storage array static.

+++ include/shared_adc.h
@@ -47,7 +47,7 @@ void initAdcModule(char adcs);
  *
  * \return The smoothed value from the ADC.
  */
-unsigned long getSmoothAdc(char adc);
+int getSmoothAdc(char adc);
 
 #endif
